requirements.txt
langchain~=0.3.11
langchain-community~=0.3.9
langchain-openai~=0.2.12
fastapi~=0.115.6
openai~=1.57.4
python-dotenv~=1.0.1
requests~=2.32.3
psycopg2-binary~=2.9.10
pydantic~=2.9.2
setuptools~=75.6.0
ipython~=8.30.0
traitlets~=5.14.3
numpy~=1.26.4
pandas~=2.2.3
SQLAlchemy~=2.0.36
decorator~=5.1.1
regex~=2024.11.6